api/requirements.txt
numpy
torch
supabase
python-dotenv
matplotlib
pillow
torchvision
sqlite3
Flask
scikit-learn
